Allow specifying additional extensions to ESLint
ESLint can be used for not just JS files, it can also be used for
TypeScript files. Currently Sputnik does not consider any other file
extension than .js for ESLint linting.

To make Sputnik pass TypeScript files to ESLint, we add an option to
consider additional file types: eslint.additionalExtensions. It expects
a comma-joined string of extensions, without "." in front. For example
"ts,tsx".

This configuration is injected into the ESLintProcessor from the
ESLintProcessorFactory. While we're at it, we also pull out
construction of the ESLintExecutor to ESLintProcessorFactory.